feat: support immutable v4
This is a breaking change that attempts to provide support
for Immutable v4 which was just released. However, the merge
function was modified to ensure that objects could not be mutated.
The tests in this repo seem to assume that mutation should be possible
so I had to update the test to reflect the new usage. This is potentially
breaking to any consumers of this repo that are relying on
that feature